% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_acf}
\alias{arima_acf}
\title{Auto-covariance function}
\usage{
arima_acf(model, n)
}
\arguments{
\item{model}{Arima model (or Sarima model)}

\item{n}{Number of auto-covariances}
}
\value{
Auto-covariances of the model. The first one corresponds to the variance and the last one to
the auto-covariance of lag n-1.
}
\description{
Auto-covariance function
}
\examples{
model<-rjd3toolkit::sarima_model(
name = "sarima",
period =12,
phi = c(-.4, .5),
d = 0,
theta = -.9,
bphi = NULL,
bd = 0,
btheta = -.7
)

ac<-arima_acf(model, 120)
}
