% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_extends_fast}
\alias{arima_extends_fast}
\title{Extends a series with ARIMA forecasts/backcasts. Exact intialization for the MA part, conditional initialization (=0)
for the AR part}
\usage{
arima_extends_fast(data, model, mean = 0, nbcasts = 0, nfcasts = 0)
}
\arguments{
\item{data}{The data to extend}

\item{model}{The Arima model}

\item{mean}{Indicates that a mean correction will be estimated}

\item{nbcasts}{Number of backcasts}

\item{nfcasts}{Number of forecasts}
}
\value{
The extended series
}
\description{
Extends a series with ARIMA forecasts/backcasts. Exact intialization for the MA part, conditional initialization (=0)
for the AR part
}
\examples{
model<-rjd3toolkit::sarima_model(
name = "sarima",
period =12,
phi = NULL,
d = 1,
theta = -.9,
bphi = NULL,
bd = 1,
btheta = -.7
)

s<-arima_extends_fast(rjd3toolkit::ABS$X0.2.09.10.M, model, nbcasts=24, nfcasts = 60)
}
