% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_pi_weights}
\alias{arima_pi_weights}
\title{Pi-weights (AR representation)}
\usage{
arima_pi_weights(model, n)
}
\arguments{
\item{model}{Arima model (or Sarima model)}

\item{n}{Number of pi-weights}
}
\value{
The n first pi-weights
}
\description{
Pi-weights (AR representation)
}
\examples{
model<-rjd3toolkit::sarima_model(
name = "sarima",
period =12,
phi = c(-.4, .5),
d = 0,
theta = -.9,
bphi = NULL,
bd = 0,
btheta = -.7
)

pi<-arima_pi_weights(model, 120)
}
