% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_psi_weights}
\alias{arima_psi_weights}
\title{Psi-weights (MA representation)}
\usage{
arima_psi_weights(model, n)
}
\arguments{
\item{model}{Arima model (or Sarima model)}

\item{n}{Number of psi-weights}
}
\value{
The n first psi-weights
}
\description{
Psi-weights (MA representation)
}
\examples{
model<-rjd3toolkit::sarima_model(
name = "sarima",
period =12,
phi = c(-.4, .5),
d = 0,
theta = -.9,
bphi = NULL,
bd = 0,
btheta = -.7
)

psi<-arima_psi_weights(model, 120)
}
