% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{bsm_outlier}
\alias{bsm_outlier}
\title{Title}
\usage{
bsm_outlier(y, bsm, X = NULL, mad = FALSE)
}
\arguments{
\item{mad}{}
}
\description{
Title
}
\examples{
bsm<-rjd3sts::bsm_model(12)
q<-bsm_outlier(rjd3toolkit::Retail$BookStores, bsm=bsm)
matplot(q[,c(4,5,6,7)], type='l')
}
