% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{bsm_outliers}
\alias{bsm_outliers}
\title{Title}
\usage{
bsm_outliers(
  y,
  bsm,
  X = NULL,
  critical.value = 0,
  ao = TRUE,
  ls = TRUE,
  so = FALSE,
  mad = TRUE,
  forward = c("Full", "Score", "Point"),
  backward = c("Point", "Score", "Full")
)
}
\arguments{
\item{backward}{}
}
\description{
Title
}
