% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdarima.R
\name{ltd_airline}
\alias{ltd_airline}
\title{Title}
\usage{
ltd_airline(
  name,
  length,
  period,
  th0 = -0.6,
  th1 = -0.6,
  bth0 = -0.6,
  bth1 = -0.6,
  fixedth = FALSE,
  variance = 0.01,
  fixedvariance = FALSE
)
}
\arguments{
\item{fixedvariance}{}
}
\description{
Title
}
