% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_regressor}
\alias{outlier_regressor}
\title{Title}
\usage{
outlier_regressor(id, period, n, pos)
}
\arguments{
\item{pos}{}
}
\description{
Title
}
\examples{
outlier_regressor("wo", 6, 120, 56)
}
