% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_regressors}
\alias{outlier_regressors}
\title{Title}
\usage{
outlier_regressors(ids, period, n, pos)
}
\arguments{
\item{pos}{}
}
\description{
Title
}
\examples{
outlier_regressors(c("wo", "ao"), 6, 120, c(56, 65))
}
