% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polynomial_at}
\alias{polynomial_at}
\title{Evaluates a polynomial for given values}
\usage{
polynomial_at(p, x)
}
\arguments{
\item{p}{Coefficients of the polynomial (for increasing powers of x, starting at the constant)}

\item{x}{Points where the polynomial must be evaluated}
}
\value{
The polynomial evaluated at the given values
}
\description{
Evaluates a polynomial for given values
}
\examples{
p<-c(1,2,3,4)
polynomial_at(p, c(1, 2))
# should be 10, 49
}
