% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polynomial_convolve}
\alias{polynomial_convolve}
\title{Convolution of a given polynomial (=P(x)*P(1/x))}
\usage{
polynomial_convolve(p)
}
\arguments{
\item{p}{The given polynomial}
}
\value{
Powers 0 to degree(p)
}
\description{
Convolution of a given polynomial (=P(x)*P(1/x))
}
\examples{
p<-c(1,2,3,4)
polynomial_convolve(p)
}
