% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polynomial_of_roots}
\alias{polynomial_of_roots}
\title{Creates a polynomial from an array of complexes (roots of the result)}
\usage{
polynomial_of_roots(roots, p0 = 1, smooth = TRUE, reorder = FALSE)
}
\arguments{
\item{roots}{Roots of the polynomial (no zero)}

\item{p0}{Value of the constant}

\item{smooth}{Smooth the coefficients (small values are set to 0)}

\item{reorder}{Reorder the roots (Leja ordering) to get more robust results}
}
\value{
The polynomial corresponding to the roots
}
\description{
Creates a polynomial from an array of complexes (roots of the result)
}
\examples{
p<-c(1, rep(0, 49), -1)
q<-polynomial_roots(p)
polynomial_of_roots(q, reorder=TRUE)
}
