% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polynomial_roots}
\alias{polynomial_roots}
\title{Computes the roots of a real polynomial}
\usage{
polynomial_roots(p, robust = FALSE)
}
\arguments{
\item{p}{Coefficients of the polynomial (for increasing powers of x, starting at the constant)}

\item{robust}{if FALSE, a "Muller-Newton" algorithm is used. Otherwise, the roots a the proper values of the companion matrix}
}
\value{
An array with the complex roots is returned
}
\description{
Computes the roots of a real polynomial
}
\examples{
p<-c(1,2,3,4)
polynomial_roots(p, TRUE)
}
