% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{regarima_outlier}
\alias{regarima_outlier}
\title{Title}
\usage{
regarima_outlier(
  y,
  arima,
  X = NULL,
  mean = FALSE,
  outliers = c("ao", "ls", "tc:0.7"),
  filter = c("fast", "fastc", "ansley", "kalman", "x12", "ljungbox"),
  mad = TRUE
)
}
\arguments{
\item{mad}{}
}
\description{
Title
}
