% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tramo_outliers}
\alias{tramo_outliers}
\title{Title}
\usage{
tramo_outliers(
  y,
  arima,
  X = NULL,
  mean = FALSE,
  critical.value = 0,
  outliers = c("ao", "ls", "tc:0.7"),
  fast = TRUE,
  ml = FALSE,
  regc = FALSE
)
}
\arguments{
\item{regc}{}
}
\description{
Title
}
