% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{x12_outliers}
\alias{x12_outliers}
\title{Title}
\usage{
x12_outliers(
  y,
  arima,
  X = NULL,
  mean = FALSE,
  critical.value = 0,
  outliers = c("ao", "ls", "tc:0.7"),
  mad = TRUE
)
}
\arguments{
\item{fast}{}
}
\description{
Title
}
